% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{add.index.lcdb}
\alias{add.index.lcdb}
\title{add a index to local database}
\usage{
add.index.lcdb(indexID)
}
\arguments{
\item{indexID}{is index code,such as EI000300}
}
\value{
nothing
}
\description{
add a index to local database
}
\examples{
add.index.lcdb(indexID="EI801003")
add.index.lcdb(indexID="EI000985")
}
\author{
Andrew Dow
}

