% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{addwgt2port_amtao}
\alias{addwgt2port_amtao}
\title{add weight to port}
\usage{
addwgt2port_amtao(port, wgtType = c("fs", "fssqrt", "ffsMV"), wgtmax = NULL,
  ...)
}
\arguments{
\item{port}{is a  object.}

\item{wgtType}{a character string, giving the weighting type of portfolio,which could be "fs"(floatingshares),"fssqrt"(sqrt of floatingshares).}

\item{wgtmax}{weight upbar.}

\item{...}{for Category Weighted method}
}
\value{
return a Port object which are of dataframe class containing at least 3 cols("date","stockID","wgt").
}
\description{
add weight to port
}
\examples{
port <- portdemo[,c('date','stockID')]
port <- addwgt2port_amtao(port)
port <- addwgt2port_amtao(port,wgtmax=0.1)
}
\author{
Andrew Dow
}
\seealso{
\code{\link[RFactorModel]{addwgt2port}}
}

