% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{bank.rotation}
\alias{bank.rotation}
\title{bank rotation}
\usage{
bank.rotation(begT, endT = Sys.Date() - 1, chgBar = 0.2)
}
\arguments{
\item{begT}{is strategy's begin date}

\item{endT}{is strategy's end date}

\item{chgBar}{is rotation's bar}
}
\value{
a list of newest bank info and strategy's historical return.
}
\description{
This is a bank stocks rotation strategy.Its idea comes from
\url{https://www.jisilu.cn/question/50176}.
}
\examples{
begT <- as.Date('2014-01-04')
endT <- Sys.Date()-1
chgBar <- 0.2
bankport <- bank.rotation(begT,endT,chgBar)
}
\author{
Andrew Dow
}

