% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{brkQT}
\alias{brkQT}
\title{bracket a series of string}
\usage{
brkQT(series)
}
\arguments{
\item{series}{is a series object}
}
\value{
a string object with bracket surrounded the series object.
}
\description{
bracket a series of string
}
\examples{
series <- c('EQ000001','EQ000002')
brkQT(series)
}
\author{
Andrew Dow
}

