% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{cbondfuns}
\alias{cbondfuns}
\alias{lcdb.build.Bond_ConBDExchangeQuote}
\alias{lcdb.update.Bond_ConBDExchangeQuote}
\title{lcdb.build.Bond_ConBDExchangeQuote}
\usage{
lcdb.build.Bond_ConBDExchangeQuote()

lcdb.update.Bond_ConBDExchangeQuote()
}
\description{
lcdb.build.Bond_ConBDExchangeQuote

lcdb.update.Bond_ConBDExchangeQuote
}
\examples{
lcdb.build.Bond_ConBDExchangeQuote()
lcdb.update.Bond_ConBDExchangeQuote()
}
