% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{db.ts}
\alias{db.ts}
\title{connect tinysoft database}
\usage{
db.ts()
}
\value{
a tinysoft conn.
}
\description{
connect tinysoft database
}
\examples{
qr <- "setsysparam(pn_stock(),'SZ000002');
setsysparam(pn_date(), today()); return nday(30,'date'
,datetimetostr(sp_time()), 'open',open(), 'close',close());"
re <- sqlQuery(db.ts(), qr);
}
\author{
Andrew Dow
}

