% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{fill.indexquote000985}
\alias{fill.indexquote000985}
\title{fill.indexquote000985}
\usage{
fill.indexquote000985()
}
\description{
fill.indexquote000985
}
\examples{
library(WindR)
w.start(showmenu = F)
fill.indexquote000985()
}

