% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{fix.lcdb.swindustry}
\alias{fix.lcdb.swindustry}
\title{fix shenwan new industry rule}
\usage{
fix.lcdb.swindustry()
}
\value{
nothing.
}
\description{
fix local database's shenwan industry rule's bug and make the rule keep consistent.
}
\examples{
fix.lcdb.swindustry()
}
\author{
Andrew Dow
}

