% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{fund_exrtn_terr}
\alias{fund_exrtn_terr}
\title{fund_exrtn_terr}
\usage{
fund_exrtn_terr(fundID, begT, endT = Sys.Date(), freq = NULL, scale = 250)
}
\arguments{
\item{fundID}{is fund ID.}

\item{begT}{is begin date.}

\item{endT}{is end date, default value is \bold{today}.}

\item{scale}{is number of periods in a year,default value is 250.}
}
\description{
calculate fund's excess return and tracking error
}
\examples{
te <- fund_exrtn_terr(fundID='162411.OF',begT=as.Date('2013-06-28'))
te <- fund_exrtn_terr(fundID=c('162411.OF','501018.OF'),begT=as.Date('2016-06-04'))
}
