% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{get_factor}
\alias{get_factor}
\alias{gf.smartQ}
\alias{gf.doublePrice}
\alias{gf.dividend}
\title{get factor}
\usage{
gf.smartQ(TS, nwin = 20, cycle = "cy_1m()", bar = 0.2)

gf.doublePrice(TS, ROEType = c("ROE_ttm", "F_ROE", "ROE", "ROE_Q"))

gf.dividend(TS, datasrc = c("jy", "wind"))
}
\arguments{
\item{TS}{is \bold{TS} object.}
}
\value{
a TSF
}
\description{
get factor
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.smartQ(TS)
TSF <- gf.dividend(TS)
TSF <- gf.doublePrice(TS)
}
