% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{getIFSpread}
\alias{getIFSpread}
\title{getIndexFuturesSpread}
\usage{
getIFSpread(begT = as.Date("2010-04-16"), endT = Sys.Date() - 1)
}
\description{
getIndexFuturesSpread
}
\examples{
re <- getIFSpread()
}
\author{
Andrew Dow
}

