% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{getIndexCompWgtNew}
\alias{getIndexCompWgtNew}
\title{getIndexCompWgtNew}
\usage{
getIndexCompWgtNew(indexID = "EI000300", endT, datasrc = defaultDataSRC())
}
\arguments{
\item{indexID}{the stockID of the index}

\item{endT}{a vector of class \code{Date}. IF missing, then get the latest components.}
}
\value{
a dataframe, with cols: "date", "stockID","wgt".
}
\description{
get the components and wgts of the specific index on certain day.
}
\examples{
tmp <- getIndexCompWgtNew("EI000300") # get the latest components
tmp <- getIndexCompWgtNew("EI000985",as.Date("2012-12-31")) # get the components on single day
tmp <- getIndexCompWgtNew("EI000985",as.Date(c("2011-12-31","2012-12-31"))) # get the components on multi-days
}

