% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{getIndexQuote_jy}
\alias{getIndexQuote_jy}
\title{get index quote from jydb}
\usage{
getIndexQuote_jy(index, begT = as.Date("2010-12-31"), endT = Sys.Date(),
  variables)
}
\arguments{
\item{index}{is a vector of charactor}

\item{begT}{is    begin date}

\item{endT}{is    end date}

\item{variables}{is a vector of charactor,such as pre_close,open,high,low,close,volume,amt,pct_chg.}
}
\value{
a data frame contains index quote.
}
\description{
get index quote from jydb
}
\examples{
index <- c('EI000300','EI000905')
begT <- as.Date("2010-12-31")
endT <- Sys.Date()
variables <- c('pct_chg')
re <- getIndexQuote_jy(index,begT,endT,variables)
}
\author{
Andrew Dow
}

