% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{getIndustryMA}
\alias{getIndustryMA}
\title{getIndustryMA}
\usage{
getIndustryMA(begT = as.Date("2005-01-04"), endT = Sys.Date() - 1)
}
\description{
getIndustryMA
}
\examples{
re <- getIndustryMA()
}
\author{
Andrew Dow
}

