% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{getIV}
\alias{getIV}
\title{getIndexValuation}
\usage{
getIV(valtype = c("PE", "PB"), caltype = c("median", "mean"),
  begT = as.Date("2005-01-04"), endT = Sys.Date() - 1)
}
\description{
getIndexValuation
}
\examples{
re <- getIV()
}
\author{
Andrew Dow
}

