% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.disposition}
\alias{gf.disposition}
\title{get disposition effect factor}
\usage{
gf.disposition(TS, nwin = 66, loadData = F)
}
\arguments{
\item{TS}{is a TS object.}

\item{nwin}{is time window.}

\item{loadData}{is logical param,for long term \bold{TS} object value should be TRUE.}
}
\value{
a TSF object
}
\description{
get disposition effect factor
}
\examples{
RebDates <- getRebDates(as.Date('2015-01-31'),as.Date('2016-9-30'),'month')
TS <- getTS(RebDates,'EI000985')
TSF <- gf.disposition(TS,loadData=T)
}
\author{
Andrew Dow
}

