% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.ILLIQ}
\alias{gf.ILLIQ}
\title{get illiquidity factor}
\usage{
gf.ILLIQ(TS, nwin = 22)
}
\arguments{
\item{TS}{is a TS object.}

\item{nwin}{is time window.}
}
\value{
a TSF object
}
\description{
get illiquidity factor
}
\examples{
RebDates <- getRebDates(as.Date('2015-01-31'),as.Date('2016-9-30'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.ILLIQ(TS)
}
\author{
Andrew Dow
}

