% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.ln_mkt_cap}
\alias{gf.ln_mkt_cap}
\title{get log(market_value) from local database}
\usage{
gf.ln_mkt_cap(TS)
}
\arguments{
\item{TS}{is a TS object.}
}
\value{
a TSF object
}
\description{
get log(market_value) from local database
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.ln_mkt_cap(TS)
}
\author{
Andrew Dow
}

