% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.volatility}
\alias{gf.volatility}
\title{get volatility factor}
\usage{
gf.volatility(TS, nwin = 60)
}
\arguments{
\item{TS}{is a TS object.}

\item{nwin}{is time window.}
}
\value{
a TSF object
}
\description{
get volatility factor
}
\examples{
RebDates <- getRebDates(as.Date('2012-01-31'),as.Date('2016-9-30'),'month')
TS <- getTS(RebDates,'EI000905')
TSF <- gf.volatility(TS)
TSF <- gf.volatility(TS,nwin=250)
}
\author{
Andrew Dow
}

