% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{gridTrade.index}
\alias{gridTrade.index}
\title{grid trading with index fund}
\usage{
gridTrade.index(indexID, begT, endT = Sys.Date() - 1, para)
}
\description{
grid trading with index fund
}
\examples{
indexID <- 'EI000905'
begT <- as.Date('2015-09-01')
endT <- Sys.Date()-1
para <- list(total=5e6,initmv=2e6,bar=0.1,mvChg=1e6,tradeCost=1/1000)
re <- gridTrade.index(indexID,begT,endT,para)
}
\author{
Andrew Dow
}

