% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{group_factor}
\alias{group_factor}
\alias{gf.SIZE}
\alias{gf.GROWTH}
\alias{gf.FORECAST}
\alias{gf.TRADING}
\alias{gf.EARNINGYIELD}
\alias{gf.VALUE}
\alias{gf.QUALITY}
\alias{gf.OTHER}
\alias{gf.LIQUIDITY}
\title{group factor}
\usage{
gf.SIZE(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.GROWTH(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.FORECAST(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.TRADING(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.EARNINGYIELD(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.VALUE(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.QUALITY(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.OTHER(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.LIQUIDITY(TS, FactorLists, wgts, factorsrc = c("local", "other"))
}
\arguments{
\item{TS}{is \bold{TS} object.}
}
\value{
a TSF
}
\description{
group factor
}
\examples{


}
