% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{index.rotation}
\alias{index.rotation}
\title{index.rotation}
\usage{
index.rotation(indexID = c("EI000016", "EI000905", "EI000852"),
  begT = as.Date("2005-01-04"), endT = Sys.Date(), MApara = 20,
  fee = 0.001, sell_count = 1)
}
\description{
index.rotation
}
\examples{
indexID <-  CT_industryList(33,1)
indexID <-  sectorID2indexID(indexID$IndustryID)
re <- index.rotation(indexID)
}
