% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{ladderNAV}
\alias{ladderNAV}
\title{ladder position port function}
\usage{
ladderNAV(assetRtn, ruledf, rebalance = NULL)
}
\arguments{
\item{assetRtn}{a data frame for stock and bond return.}

\item{ruledf}{a data frame for position rule.}

\item{rebalance}{rebalance frequency, default value is NULL.}
}
\value{
a data frame, containing return,postion,nav.
}
\description{
ladder position port function
}
\examples{
assetRtn <- rtndemo
ruledf <- data.frame(nav = c(0,1.05,1.10,1.20),pos = c(0.15,0.25,0.35,0.5))
re <- ladderNAV(assetRtn,ruledf)
rebalance <- '2 years'
re <- ladderNAV(assetRtn,ruledf,rebalance)
}
\author{
han.qian
}

