% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{lcdb.add.QT_IndexQuote}
\alias{lcdb.add.QT_IndexQuote}
\title{lcdb.add.QT_IndexQuote}
\usage{
lcdb.add.QT_IndexQuote(indexID)
}
\description{
lcdb.add.QT_IndexQuote
}
\examples{
lcdb.add.QT_IndexQuote("EI000852")
}
