% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{lcdb.init.IndexQuote_000985E}
\alias{lcdb.init.IndexQuote_000985E}
\title{lcdb.init.IndexQuote_000985E}
\usage{
lcdb.init.IndexQuote_000985E()
}
\description{
lcdb.init.IndexQuote_000985E
}
\examples{
lcdb.init.IndexQuote_000985E()
}

