% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{lcdb.update.CorpStockPool}
\alias{lcdb.update.CorpStockPool}
\title{lcdb.update.CorpStockPool}
\usage{
lcdb.update.CorpStockPool(filenames)
}
\arguments{
\item{filenames}{a vector of filename with path.}
}
\description{
lcdb.update.CorpStockPool
}
\examples{
filenames <- c('D:/sqlitedb/core.csv','D:/sqlitedb/preclose.csv')
lcdb.update.CorpStockPool(filenames)
}

