% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{lcdb.update.QT_IndexQuoteamtao}
\alias{lcdb.update.QT_IndexQuoteamtao}
\title{lcdb.update.QT_IndexQuoteamtao}
\usage{
lcdb.update.QT_IndexQuoteamtao(begT, endT, IndexID, datasrc = "quant")
}
\description{
lcdb.update.QT_IndexQuoteamtao
}
