% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{lcdb.update.QT_IndexTiming}
\alias{lcdb.update.QT_IndexTiming}
\title{update table QT_IndexTiming}
\usage{
lcdb.update.QT_IndexTiming()
}
\description{
update table QT_IndexTiming
}
\examples{
lcdb.update.QT_IndexTiming()
}
\author{
Andrew Dow
}

