% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{lcdb.update.QT_IndexValuation}
\alias{lcdb.update.QT_IndexValuation}
\title{lcdb.update.QT_IndexValuation}
\usage{
lcdb.update.QT_IndexValuation()
}
\description{
lcdb.update.QT_IndexValuation
}
\examples{
lcdb.update.QT_IndexValuation()
}
\author{
Andrew Dow
}
