% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerFun.R
\name{lcfs.update.amtao}
\alias{lcfs.update.amtao}
\title{lcfs.update.amtao}
\usage{
lcfs.update.amtao(factorFun, factorPar, factorDir, factorID, begT, endT,
  splitNbin)
}
\description{
lcfs.update.amtao
}
\examples{

factorFun="gf.ILLIQ"
factorPar=""
factorDir=1
factorID="F000018"
begT = as.Date("2000-01-01")
endT = Sys.Date()-1
splitNbin = "month"
lcfs.update.amtao()
}

