% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{LLT}
\alias{LLT}
\title{LLT timing}
\usage{
LLT(indexID = "EI000300", begT = as.Date("2005-01-04"), d = 60,
  trancost = 0.001, type = c("LLT", "SMA"))
}
\description{
LLT timing
}
\examples{
re <- LLT()
}
\author{
Andrew Dow
}

