% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{POFund}
\alias{POFund}
\title{private offering fund}
\usage{
POFund(fundID, begT, endT)
}
\description{
get private offering fund daily nav and premium and discount ration info
}
\examples{
re <- POFund(fundID,begT,endT)
}
\author{
Andrew Dow
}

