% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\docType{data}
\name{portdemo}
\alias{portdemo}
\title{alpha portfolio demo data}
\format{A data frame with 1149 rows and 3 variables.}
\usage{
portdemo
}
\description{
part of index EI000905's alpha portfolio data set.
}
\keyword{datasets}

