% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{pure.factor.test}
\alias{pure.factor.test}
\title{get purified factor's portfolio}
\usage{
pure.factor.test(TSFR, riskfactorLists)
}
\arguments{
\item{TSFR}{is a TSFR objetc,factorscore should be standardized and remove NA values}

\item{riskfactorLists}{is a list of risky factor,such as market value,beta,etc.}
}
\value{
a list of pure factor portfolio's return and weight.
}
\description{
remove style and industry risky factor,get the pure alpha factor's return
}
\examples{
begT <- as.Date('2012-01-31')
endT <- as.Date('2016-06-30')
RebDates <- getRebDates(begT,endT)
TS <- getTS(RebDates,'EI000905')
TSF <- getTSF(TS,factorFun = 'gf_lcfs',factorPar = list(factorID='F000017'),factorDir = -1,
              factorStd = 'norm')
TSFR <- getTSR(TSF)
riskfactorLists <- buildFactorLists(buildFactorList(factorFun = "gf.ln_mkt_cap",factorDir = -1,
                         factorStd = "norm"))
factorIDs <- c("F000006","F000015","F000016")
tmp <- buildFactorLists_lcfs(factorIDs,factorStd="norm")
riskfactorLists <- c(riskfactorLists,tmp)
purefp <- pure.factor.test(TSFR,riskfactorLists)
}
\author{
Andrew Dow
}

