% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{refinePar_zz}
\alias{refinePar_zz}
\title{refinePar_zz}
\usage{
refinePar_zz(type = c("zz", "none"), sectorAttr = defaultSectorAttr(),
  log = FALSE, regLists = list(fl_cap(log = TRUE)))
}
\description{
refinePar_zz
}
\examples{
refinePar <- refinePar_zz()
}
