% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{resumeArbitrage}
\alias{resumeArbitrage}
\title{resumption stock arbitrage}
\usage{
resumeArbitrage(begT, endT)
}
\description{
resumption stock arbitrage
}
\examples{
begT <- Sys.Date()-1
endT <- Sys.Date()
re <- resumeArbitrage(begT,endT)
}
\author{
Andrew Dow
}

