% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{rmPriceLimit}
\alias{rmPriceLimit}
\title{remove price limits}
\usage{
rmPriceLimit(TS, dateType = c("nextday", "today"), priceType = c("upLimit",
  "downLimit"))
}
\description{
remove price limits
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
TSnew <- rmPriceLimit(TS,dateType='today',priceType='downLimit')
}
\author{
Andrew Dow
}

