% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{rmSuspend}
\alias{rmSuspend}
\title{remove suspension stock from TS}
\usage{
rmSuspend(TS)
}
\arguments{
\item{TS}{is a \bold{TS} object.}
}
\value{
return a \bold{TS} object.
}
\description{
remove suspension stock from TS
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
TS <- rmSuspend(TS)
}
\author{
Andrew Dow
}

