% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{rtn.persum}
\alias{rtn.persum}
\title{combine rtn.periods and rtn.summary}
\usage{
rtn.persum(rtn, freq = "year", Rf = 0, showPer = T)
}
\arguments{
\item{rtn}{an xts, vector, matrix, data frame, timeSeries or zoo object of asset returns}

\item{freq}{An interval specification, one of "day", "week", "month", "quarter" and "year", optionally preceded by an integer and a space, or followed by "s".See \code{\link{cut.Date}} for detail.}

\item{Rf}{risk free rate, in same period as your returns}
}
\value{
a matrix, giving the summary infomation of the rtn series,including Annualized Return,Annualized Std Dev,Annualized Sharpe,HitRatio,Worst Drawdown
}
\description{
combine rtn.periods and rtn.summary
}
\examples{
rtn <- rtndemo
rtn <- xts::xts(rtn[,-1],rtn[,1])
rtn.persum(rtn)
}
\seealso{
\code{\link[QUtility]{rtn.periods}}

\code{\link[QUtility]{rtn.summary}}
}

