% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\docType{data}
\name{rtndemo}
\alias{rtndemo}
\title{assets return demo dataset.}
\format{A data frame with 1627 rows and 3 variables:
\describe{
  \item{date}{date type}
  \item{stock}{stock index return}
  \item{bond}{bond index return}
}}
\usage{
rtndemo
}
\description{
A dataset containing stock index and bond index daily return data since 2009.
}
\keyword{datasets}

