% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{table.factor.summary}
\alias{table.factor.summary}
\title{combine funcs table.IC and table.Ngroup.spread}
\usage{
table.factor.summary(TSFR, N = 10, fee = 0.001)
}
\arguments{
\item{TSFR}{is a TSFR object}

\item{N}{seprate the stocks into N group}

\item{fee}{trading cost}
}
\value{
a factor's IC summary and longshort portfolio's return summary.
}
\description{
combine funcs table.IC and table.Ngroup.spread
}
\examples{
modelPar <- modelPar.default()
TSFR <- Model.TSFR(modelPar)
table.factor.summary(TSFR)
}
\author{
Andrew Dow
}

