% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{TS.getFinStat_ts}
\alias{TS.getFinStat_ts}
\title{TS.getFinStat_ts}
\usage{
TS.getFinStat_ts(TS, funchar, varname = funchar,
  Nbin = lubridate::years(-3), growth = FALSE, stattype = c("mean",
  "slope", "sd", "mean/sd", "slope/sd"))
}
\arguments{
\item{funchar}{see \link[QDataGet]{TS.getFin_rptTS}}

\item{growth}{whether get the growing rate of financial indicators,default value is \code{TRUE}.}
}
\description{
get stats of financial indicators from tinysoft.
}
\examples{
RebDates <- getRebDates(as.Date('2013-01-31'),as.Date('2017-05-31'),'month')
TS <- getTS(RebDates,'EI000905')
funchar <- 'LastQuarterData(RDate,46078,0)'
varname <- 'np'
TSF <- TS.getFinStat_ts(TS,funchar,varname,growth=TRUE)
funchar <- 'LastQuarterData(RDate,9900000,0)'
varname <- 'eps'
TSF <- TS.getFinStat_ts(TS,funchar,varname)
}
