% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{TSFR.rptTSF_nextF}
\alias{TSFR.rptTSF_nextF}
\title{TSFR.rptTSF_nextF}
\usage{
TSFR.rptTSF_nextF(begT = as.Date("2007-12-31"),
  endT = as.Date("2016-12-31"), freq = "y", univ, funchar,
  refinePar = refinePar_default(), ...)
}
\arguments{
\item{funchar}{can be missing when ... arguments are passed.}

\item{...}{argument for \code{\link[QDataGet]{getTSF}}.}
}
\value{
a \bold{TSFR} object,\code{date_end} represents next report date,\code{factorscore} represents factorscore of next report date.
}
\description{
TSFR.rptTSF_nextF
}
\examples{
begT <- as.Date("2007-12-31")
endT <- as.Date("2017-12-31")
freq <- "y"
univ <- "EI000985"
funchar <- '"factorscore",reportofall(9900100,RDate)' #ROE
funchar <- '"factorscore",reportofall(9900604,RDate)' #growth of net profit
funchar <- '"factorscore",reportofall(9900501,RDate)' #divdendyield
refinePar <- refinePar_default(type="scale",sectorAttr=NULL)
refinePar <- refinePar_default(type="scale",sectorAttr=defaultSectorAttr())
TSFR <- TSFR.rptTSF_nextF(begT,endT,freq,univ,funchar,refinePar)
TSFR <- TSFR.rptTSF_nextF(begT,endT,freq,univ,refinePar=refinePar,factorFun='gf.PB_mrq', factorPar = list(fillna = TRUE))
}
