% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{update.lcdb.RegTables}
\alias{update.lcdb.RegTables}
\title{update local database's regression tables}
\usage{
\method{update}{lcdb.RegTables}()
}
\description{
update local database's regression tables
}

