% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{bank.rotation}
\alias{bank.rotation}
\title{bank rotation}
\usage{
bank.rotation(begDate, endDate = Sys.Date(), chgBar = 0.2)
}
\arguments{
\item{begDate}{is strategy's begin date}

\item{endDate}{is strategy's end date}

\item{chgBar}{is rotation's bar}
}
\value{
a list of newest bank info and strategy's historical return.
}
\description{
This is a bank stocks rotation strategy.Its idea comes from
\url{https://www.jisilu.cn/question/50176}.
}
\examples{
begDate <- as.Date('2014-01-04')
endDate <- Sys.Date()-1
chgBar <- 0.2
bankport <- bank.rotation(begDate,endDate,chgBar)
}
\author{
Andrew Dow
}

