% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{get_factor}
\alias{get_factor}
\alias{gf.smartQ}
\alias{gf.doublePrice}
\alias{gf.dividendyield}
\alias{gf.BP_mrq}
\alias{gf.EP_ttm}
\alias{gf.CFP_ttm}
\alias{gf.ROA_ttm}
\alias{gf.momentum}
\alias{gf.lev_dtoa}
\alias{gf.lev_ldtoop}
\alias{gf.nl_size}
\alias{gf.pio_f_score}
\title{get factor}
\usage{
gf.smartQ(TS, nwin = 20, cycle = "cy_1m()", bar = 0.2)

gf.doublePrice(TS, ROEType = c("ROE_ttm", "F_ROE", "ROE", "ROE_Q"))

gf.dividendyield(TS, datasrc = c("ts", "jy", "wind"))

gf.BP_mrq(TS)

gf.EP_ttm(TS)

gf.CFP_ttm(TS)

gf.ROA_ttm(TS)

gf.momentum(TS, Nlong = 250, Nshort = 20)

gf.lev_dtoa(TS)

gf.lev_ldtoop(TS)

gf.nl_size(TS)

gf.pio_f_score(TS)
}
\arguments{
\item{TS}{is \bold{TS} object.}
}
\value{
a TSF
}
\description{
get factor
}
\examples{
RebDates <- getRebDates(as.Date('2011-03-17'),as.Date('2012-04-17'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.smartQ(TS)
TSF <- gf.dividendyield(TS)
TSF <- gf.doublePrice(TS)
}
\author{
han.qian
}
