% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{getIndexBasicInfo}
\alias{getIndexBasicInfo}
\title{getIndexBasicInfo}
\usage{
getIndexBasicInfo(indexID)
}
\arguments{
\item{indexID}{is set of index ID.}
}
\description{
getIndexBasicInfo
}
\examples{
index <- getIndexBasicInfo('EI000300')
index <- getIndexBasicInfo(c('EI000300','EI000905'))
}
