% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.beta}
\alias{gf.beta}
\title{get beta factor}
\usage{
gf.beta(TS)
}
\arguments{
\item{TS}{is a TS object.}
}
\value{
a TSF object
}
\description{
get beta factor
}
\examples{
RebDates <- getRebDates(as.Date('2015-01-31'),as.Date('2015-12-31'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.beta(TS)
}
\author{
Andrew Dow
}

