% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{gf.IVR}
\alias{gf.IVR}
\title{get IVR factor}
\usage{
gf.IVR(TS, nwin = 22)
}
\arguments{
\item{TS}{is a TS object.}

\item{nwin}{time window}
}
\value{
a TSF object
}
\description{
get IVR factor
}
\examples{
RebDates <- getRebDates(as.Date('2015-06-30'),as.Date('2015-12-31'),'month')
TS <- getTS(RebDates,'EI000300')
TSF <- gf.IVR(TS)
}
\author{
Andrew Dow
}

