% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{group_factor}
\alias{group_factor}
\alias{gf.SIZE}
\alias{gf.GROWTH}
\alias{gf.TRADING}
\alias{gf.EARNINGYIELD}
\alias{gf.VALUE}
\alias{gf.OTHER}
\title{group factor}
\usage{
gf.SIZE(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)

gf.GROWTH(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)

gf.TRADING(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)

gf.EARNINGYIELD(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)

gf.VALUE(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)

gf.OTHER(TS, refinePar = refinePar_default("scale"), FactorLists, wgt)
}
\arguments{
\item{TS}{is \bold{TS} object.}
}
\value{
a TSF
}
\description{
group factor
}
\examples{


}
