% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{group_factor}
\alias{group_factor}
\alias{gf.group_liquidity}
\alias{gf.group_size}
\alias{gf.group_growth}
\alias{gf.group_forecast}
\alias{gf.group_trading}
\alias{gf.group_earningyield}
\alias{gf.group_value}
\alias{gf.group_quality}
\alias{gf.group_other}
\title{group factor}
\usage{
gf.group_liquidity(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_size(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_growth(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_forecast(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_trading(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_earningyield(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_value(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_quality(TS, FactorLists, wgts, factorsrc = c("local", "other"))

gf.group_other(TS, FactorLists, wgts, factorsrc = c("local", "other"))
}
\arguments{
\item{TS}{is \bold{TS} object.}
}
\value{
a TSF
}
\description{
group factor
}
\examples{


}
