% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{index_valuation}
\alias{index_valuation}
\alias{lcdb.build.QT_IndexValuation}
\alias{lcdb.addindex.QT_IndexValuation}
\alias{lcdb.update.QT_IndexValuation}
\alias{getIV}
\title{lcdb.build.QT_IndexValuation}
\usage{
lcdb.build.QT_IndexValuation()

lcdb.addindex.QT_IndexValuation(indexset)

lcdb.update.QT_IndexValuation()

getIV(valtype = c("PE", "PB"), caltype = c("median", "mean"),
  begT = as.Date("2005-01-04"), endT = Sys.Date() - 1)
}
\description{
lcdb.build.QT_IndexValuation

lcdb.addindex.QT_IndexValuation

lcdb.update.QT_IndexValuation

getIndexValuation
}
\examples{
lcdb.build.QT_IndexValuation()
lcdb.addindex.QT_IndexValuation()
lcdb.update.QT_IndexValuation()
re <- getIV()
}
\author{
Andrew Dow

Andrew Dow
}
