% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{index_valuation}
\alias{index_valuation}
\alias{lcdb.build.QT_IndexValuation}
\alias{lcdb.update.QT_IndexValuation}
\alias{getIndexValuation}
\title{lcdb.build.QT_IndexValuation}
\usage{
lcdb.build.QT_IndexValuation(indexID = c("EI399006", "EI000933"),
  addIndex = FALSE)

lcdb.update.QT_IndexValuation(begT, endT = Sys.Date() - 1)

getIndexValuation(valtype = c("PE", "PB"), caltype = c("median", "mean"),
  begT = as.Date("2005-01-04"), endT = Sys.Date() - 1)
}
\description{
lcdb.build.QT_IndexValuation

lcdb.update.QT_IndexValuation
}
\examples{
lcdb.build.QT_IndexValuation()
lcdb.update.QT_IndexValuation()
re <- getIndexValuation()
re <- getIndexValuation(caltype='median',begT = Sys.Date(),endT = Sys.Date())
}
\author{
Andrew Dow
}
