% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{risk.parity}
\alias{risk.parity}
\title{risk parity}
\usage{
risk.parity(asset, rebFreq = "month", training = 250)
}
\arguments{
\item{asset}{is \bold{\link{xts}} object.}

\item{begT}{is begin date.}

\item{endT}{is end date, default value is \bold{today}.}
}
\description{
risk parity
}
\examples{
suppressMessages(library(PortfolioAnalytics))
asset <- rtndemo
asset <- xts::xts(asset[,-1],order.by = asset[,1])


}

