% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\name{risk.parity}
\alias{risk.parity}
\title{risk parity}
\usage{
risk.parity(asset, rebFreq = "month", training = 250, riskcontri)
}
\arguments{
\item{fundID}{is fund ID.}

\item{begT}{is begin date.}

\item{endT}{is end date, default value is \bold{today}.}

\item{scale}{is number of periods in a year,default value is 250.}
}
\description{
risk parity
}
\examples{
# get data from wind
library(WindR)
w.start(showmenu = F)
te <- fundTE(fundID='162411.OF',begT=as.Date('2013-06-28'))
te <- fundTE(fundID='501018.OF',begT=as.Date('2016-06-28'))

# get data from jy

}

