% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{rmNegativeEvents}
\alias{rmNegativeEvents}
\title{remove  from TS}
\usage{
rmNegativeEvents(TS, type = c("AnalystDown", "ForcastReportLoss",
  "ShareholderReduction", "all"))
}
\arguments{
\item{TS}{is a \bold{TS} object.}

\item{type}{is negative events' type.}
}
\value{
return a \bold{TS} object.
}
\description{
remove  from TS
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
TS <- rmNegativeEvents(TS)
}
\author{
Andrew Dow
}

