% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorModelPart.R
\name{rmSuspend}
\alias{rmSuspend}
\title{remove suspension stock from TS}
\usage{
rmSuspend(TS, type = c("nextday", "today", "both"),
  datasrc = defaultDataSRC())
}
\arguments{
\item{TS}{is a \bold{TS} object.}

\item{type}{is suspension type.}
}
\value{
return a \bold{TS} object.
}
\description{
remove suspension stock from TS
}
\examples{
RebDates <- getRebDates(as.Date('2013-03-17'),as.Date('2016-04-17'),'month')
TS <- getTS(RebDates,'EI000985')
TSnew <- rmSuspend(TS,type='today')
}
\author{
Andrew Dow
}

