% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trickFun.R
\docType{data}
\name{rtndemo}
\alias{rtndemo}
\title{assets return demo dataset.}
\format{A data frame with 2865 rows and 4 variables:
\describe{
  \item{date}{date type}
  \item{stock}{stock index return}
  \item{bond}{bond index return}
  \item{commodity}{commodity index return}
}}
\usage{
rtndemo
}
\description{
A dataset containing stock index(000985.CSI), bond index(037.CS) and commodity(GC00.CMX) daily return data since 2009.
}
\keyword{datasets}

